inlined stuff in CopyTo2x2()
